backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
